refactor(stress): Rename inner module to `stress`, drop custom #[path]
