+++ mountpoint-s3-fs/tests/stress/harness/invariants.rs
