+++ mountpoint-s3-fs/tests/stress/harness/memory_monitor.rs
