+++ mountpoint-s3-fs/tests/stress/harness/scenario.rs
