+++ mountpoint-s3-fs/tests/stress/harness/watchdog.rs
