+++ mountpoint-s3-fs/tests/stress/harness/worker.rs
