+++ mountpoint-s3-fs/tests/stress/mod.rs
@@ -1,0 +1,6 @@
+//! Stress-test harness, workers, scenarios, and shared test objects.
+
+pub mod harness;
+pub mod scenarios;
+pub mod test_objects;
+pub mod workers;
