+++ mountpoint-s3-fs/tests/stress/test_objects.rs
