+++ mountpoint-s3-fs/tests/stress/workers/idle.rs
@@ -6,8 +6,8 @@ use std::path::Path;
 use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
 use std::time::{Duration, Instant};
 
-use crate::harness::{FileOp, FileOpLatencies, Worker};
-use crate::test_objects::SHARED_OBJECTS_PREFIX;
+use crate::stress::harness::{FileOp, FileOpLatencies, Worker};
+use crate::stress::test_objects::SHARED_OBJECTS_PREFIX;
 
 use super::churn::pick_index;
 use super::common::SharedObjectPool;
