+++ mountpoint-s3-fs/tests/stress/workers/sequential_reader.rs
@@ -3,8 +3,8 @@
 use std::path::Path;
 use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
 
-use crate::harness::{FileOpLatencies, Worker};
-use crate::test_objects::SHARED_OBJECTS_PREFIX;
+use crate::stress::harness::{FileOpLatencies, Worker};
+use crate::stress::test_objects::SHARED_OBJECTS_PREFIX;
 
 use super::common::{SharedObject, read_to_eof_once};
